Cleaned up eosio helper functions
