+++ src/index.ts
@@ -2,6 +2,10 @@ export * from './authenticator';
 
 export * from './user';
 
-export { createDelegatedAuthority, createKeyAuthoriy, addCodePermission } from './services/eosio';
+import * as Authority from './services/eosio/authority';
+import * as Eosio from './services/eosio/eosio';
+import * as Transaction from './services/eosio/transaction';
+const EosioUtil = { ...Authority, ...Eosio, ...Transaction };
+export { EosioUtil };
 
 export * as Services from './services/IDContract';
