+++ src/services/eosio/authority.ts
@@ -1,14 +1,3 @@
-import { APIClient, FetchProvider, PrivateKey } from "@greymass/eosio";
-import fetch from 'cross-fetch';
-
-const privateKey = PrivateKey.from("PVT_K1_2bfGi9rYsXQSXXTvJbDAPhHLQUojjaNLomdm3cEJ1XTzMqUt3V");
-const publicKey = privateKey.toPublic();
-// PUB_K1_6MRyAjQq8ud7hVNYcfnVPJqcVpscN5So8BhtHuGYqET5BoDq63
-
-const api = new APIClient({
-    url: "http://localhost:8888",
-    provider: new FetchProvider("http://localhost:8888", { fetch })
-})
 
 function createKeyAuthoriy(key: string) {
     return {
@@ -46,4 +35,4 @@ function addCodePermission(authority: any, account: string) {
     return authority;
 }
 
-export { api, privateKey, publicKey, createDelegatedAuthority, createKeyAuthoriy, addCodePermission };
+export { createDelegatedAuthority, createKeyAuthoriy, addCodePermission };
